+++ Samples/WinCopies.GUI.Samples.Shared/ExplorerControlWindow.cs
@@ -1,10 +1,36 @@
-﻿using WinCopies.GUI.Shell;
+﻿/* Copyright © Pierre Sprimont, 2021
+ *
+ * This file is part of the WinCopies Framework.
+ *
+ * The WinCopies Framework is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * The WinCopies Framework is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with the WinCopies Framework.  If not, see <https://www.gnu.org/licenses/>. */
+
+using WinCopies.GUI.Shell;
 
 namespace WinCopies.GUI.Samples
 {
     public partial class ExplorerControlWindow : BrowsableObjectInfoWindow
     {
-        public ExplorerControlWindow() : base(new BrowsableObjectInfoWindowViewModel()) { /* Left empty. */ }
+        public ExplorerControlWindow() : base(GetDefaultDataContext()) { /* Left empty. */ }
+
+        public static BrowsableObjectInfoWindowViewModel GetDefaultDataContext()
+        {
+            var dataContext = new BrowsableObjectInfoWindowViewModel();
+
+            dataContext.Paths.IsCheckBoxVisible = true;
+
+            return dataContext;
+        }
 
         protected override BrowsableObjectInfoWindow GetNewBrowsableObjectInfoWindow() => new ExplorerControlWindow();