+++ Source/IO/WinCopies.GUI.Shell.Shared/BrowsableObjectInfoWindow.cs
@@ -21,6 +21,7 @@ using Microsoft.WindowsAPICodePack.Shell;
 
 using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Reflection;
 using System.Windows;
@@ -71,7 +72,7 @@ namespace WinCopies.GUI.Shell
             {
                 var listView = (ExplorerControlListView)e.OriginalSource;
 
-                BrowsableObjectInfoWindow window = listView.GetParent<BrowsableObjectInfoWindow>(true);
+                BrowsableObjectInfoWindow window = listView.GetParent<BrowsableObjectInfoWindow>(false);
 
                 IExplorerControlBrowsableObjectInfoViewModel selectedItem = ((BrowsableObjectInfoWindowViewModel)window.DataContext).Paths.SelectedItem;
 
@@ -115,8 +116,6 @@ namespace WinCopies.GUI.Shell
                 dataContext.Paths.Paths.Add(GetDefaultExplorerControlBrowsableObjectInfoViewModel());
 
             AddDefaultCommandBindings();
-
-            // InitializeComponent();
         }
 
         private void Command_CanExecute(object sender, CanExecuteRoutedEventArgs e)
@@ -128,7 +127,7 @@ namespace WinCopies.GUI.Shell
 
         private void CloseAllTabs_Executed(object sender, ExecutedRoutedEventArgs e)
         {
-            System.Collections.ObjectModel.ObservableCollection<IExplorerControlBrowsableObjectInfoViewModel> paths = ((BrowsableObjectInfoWindowViewModel)DataContext).Paths.Paths;
+            ObservableCollection<IExplorerControlBrowsableObjectInfoViewModel> paths = ((BrowsableObjectInfoWindowViewModel)DataContext).Paths.Paths;
 
             paths.Clear();
 
@@ -141,13 +140,19 @@ namespace WinCopies.GUI.Shell
             e.Handled = true;
         }
 
-        protected virtual bool OnCanCancelClose(System.Collections.ObjectModel.ObservableCollection<IExplorerControlBrowsableObjectInfoViewModel> paths) => /* !Current.IsClosing && */ paths.Count > 1 && MessageBox.Show(this, Shell.Properties.Resources.WindowClosingMessage, "WinCopies", YesNo, Question, No) != Yes;
+        protected virtual bool OnCanCancelClose(ObservableCollection<IExplorerControlBrowsableObjectInfoViewModel> paths) => /* !Current.IsClosing && */ paths.Count > 1 && MessageBox.Show(this, Properties.Resources.WindowClosingMessage, "WinCopies", YesNo, Question, No) != Yes;
 
         protected override void OnClosing(CancelEventArgs e)
         {
-            if (OnCanCancelClose(((BrowsableObjectInfoWindowViewModel)DataContext).Paths.Paths))
+            ObservableCollection<IExplorerControlBrowsableObjectInfoViewModel> paths = ((BrowsableObjectInfoWindowViewModel)DataContext).Paths.Paths;
+
+            if (OnCanCancelClose(paths))
 
                 e.Cancel = true;
+
+            else
+
+                paths.Clear();
 
             base.OnClosing(e);
         }